+++ Changelog.md
@@ -62,6 +62,10 @@
   * Added `DRT_SAVE_SNAPSHOTS` which enables saving snapshots of the layout each
     detalied routing iteration.
   * Added `DRT_SAVE_DRC_REPORT_ITERS`
+  * Added `DRT_ANTENNA_REPAIR_ITERS`, which if greater than zero, enables
+    antenna fixing after detailed routing
+  * Added `DRT_ANTENNA_MARGIN` which is similar to `GRT_ANTENNA_MARGIN` but for
+    the aforementioned antenna repair iterations
 
 * `OpenROAD.GlobalPlacement`